requirements.txt
absl-py==0.6.1
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.5
colorama==0.3.9
conda==4.5.4
cryptography==2.2.2
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.1.1
menuinst==1.4.14
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.6.0
numpy==1.15.4
opencv-python==3.4.2.17
pandas==0.22.0
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.4
Pillow==5.0.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
pycosat==0.6.3
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.0
requests==2.18.4
ruamel-yaml==0.15.37
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.0.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.22
vboxapi==1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
win-inet-pton==1.0.1
wincertstore==0.2
